pyproject.toml
[project]
name = "project-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jax>=0.8.2",
    "matplotlib>=3.10.8",
    "pyyaml>=6.0.3",
]
